Train yolov3 and Eval yolovx

+++ scripts/model/utils/utils.py
@@ -457,10 +457,12 @@ def build_targets(p, targets, model):
         gwh = t[:, 4:6]  # grid wh
         gij = (gxy - offsets).long()
         gi, gj = gij.T  # grid xy indices
-
+        gi=gi.float()
+        gj=gj.float()
         # Append
         a = t[:, 6].long()  # anchor indices
-        indices.append((b, a, gj.clamp_(0, gain[3] - 1), gi.clamp_(0, gain[2] - 1)))  # image, anchor, grid indices
+        # print(type(b), type(a), type(gain[3]))
+        indices.append((b, a, gj.clamp_(0, gain[3].float() - 1).long(), gi.clamp_(0, gain[2].float()- 1).long()))  # image, anchor, grid indices
         tbox.append(torch.cat((gxy - gij, gwh), 1))  # box
         anch.append(anchors[a])  # anchors
         tcls.append(c)  # class
